requirements-dev.txt
-r requirements.txt

factory-boy==3.3.3
httpx==0.28.1
mypy==1.16.1
pytest==8.4.1
pytest-asyncio==1.0.0
ruff==0.12.0
types-factory-boy==0.4.1
